Q2/requirements.txt
langchain
langchain_core
langchain_openai 
pandas
gspread 
gspread-dataframe 
google-auth-oauthlib 
langchain-google-genai 
langchain-experimental
python-dotenv
ruff
tabulate
streamlit